Revert "test: move placeholder img util to tests and use in list"
This reverts commit 8cd6a6f7b4f6e1836ebc8e22912c0c54a93870a6.
