+++ src/tests/utils.ts
@@ -271,78 +271,3 @@ export async function skipAnimations(page: E2EPage): Promise<void> {
     content: `:root { --calcite-duration-factor: 0; }`
   });
 }
-
-/*
-MIT License
-
-Copyright (c) 2020 Cloud Four
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-https://github.com/cloudfour/simple-svg-placeholder
-*/
-
-interface SimpleSvgPlaceholderParams {
-  width?: number;
-  height?: number;
-  text?: string;
-  fontFamily?: string;
-  fontWeight?: string;
-  fontSize?: number;
-  dy?: number;
-  bgColor?: string;
-  textColor?: string;
-  dataUri?: boolean;
-  charset?: string;
-}
-
-export function placeholderImage({
-  width = 300,
-  height = 150,
-  text = `${width}×${height}`,
-  fontFamily = "sans-serif",
-  fontWeight = "bold",
-  fontSize = Math.floor(Math.min(width, height) * 0.2),
-  dy = fontSize * 0.35,
-  bgColor = "#ddd",
-  textColor = "rgba(0,0,0,0.5)",
-  dataUri = true,
-  charset = "UTF-8"
-}: SimpleSvgPlaceholderParams): string {
-  const str = `<svg xmlns="http://www.w3.org/2000/svg" width="${width}" height="${height}" viewBox="0 0 ${width} ${height}">
-    <rect fill="${bgColor}" width="${width}" height="${height}"/>
-    <text fill="${textColor}" font-family="${fontFamily}" font-size="${fontSize}" dy="${dy}" font-weight="${fontWeight}" x="50%" y="50%" text-anchor="middle">${text}</text>
-  </svg>`;
-
-  // Thanks to: filamentgroup/directory-encoder
-  const cleaned = str
-    .replace(/[\t\n\r]/gim, "") // Strip newlines and tabs
-    .replace(/\s\s+/g, " ") // Condense multiple spaces
-    .replace(/'/gim, "\\i"); // Normalize quotes
-
-  if (dataUri) {
-    const encoded = encodeURIComponent(cleaned)
-      .replace(/\(/g, "%28") // Encode brackets
-      .replace(/\)/g, "%29");
-
-    return `data:image/svg+xml;charset=${charset},${encoded}`;
-  }
-
-  return cleaned;
-}
